+++ config.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/go-viper/mapstructure/v2"
 
+	"github.com/nil-go/konf/internal/credential"
 	"github.com/nil-go/konf/internal/maps"
 )
 
@@ -82,10 +83,11 @@ func New(opts ...Option) Config {
 // Each loader takes precedence over the loaders before it.
 //
 // This method can be called multiple times but it is not concurrency-safe.
-// It panics if loader is nil.
 func (c Config) Load(loader Loader, opts ...LoadOption) error {
 	if loader == nil {
-		panic("cannot load config from nil loader")
+		c.logger.Warn("cannot load config from nil loader")
+
+		return nil
 	}
 
 	loadOption := &loadOptions{}
@@ -134,62 +136,27 @@ func (c Config) Unmarshal(path string, target any) error {
 		return fmt.Errorf("new decoder: %w", err)
 	}
 
-	if err := decoder.Decode(sub(c.values.values, strings.Split(path, c.delimiter))); err != nil {
+	if err := decoder.Decode(maps.Sub(c.values.values, strings.Split(path, c.delimiter))); err != nil {
 		return fmt.Errorf("decode: %w", err)
 	}
 
 	return nil
 }
 
-func sub(values map[string]any, keys []string) any {
-	if len(keys) == 0 || len(keys) == 1 && keys[0] == "" {
-		return values
-	}
-
-	var next any = values
-	for _, key := range keys {
-		key = strings.ToLower(key)
-		mp, ok := next.(map[string]any)
-		if !ok {
-			return nil
-		}
-
-		val, exist := mp[key]
-		if !exist {
-			return nil
-		}
-		next = val
-	}
-
-	return next
-}
-
 // Explain provides information about how Config resolve each value
-// from loaders for the given path.
+// from loaders for the given path. It blur sensitive information.
 // The path is case-insensitive.
-//
-// If there are sensitive information (e.g. password, secret) which should not be exposed,
-// you can use [WithValueFormatter] to pass a value formatter to blur the information.
-// By default, it uses CredentialFormatter to blur sensitive information.
-func (c Config) Explain(path string, opts ...ExplainOption) string {
-	option := &explainOptions{}
-	for _, opt := range opts {
-		opt(option)
-	}
-	if option.valueFormatter == nil {
-		option.valueFormatter = CredentialFormatter
-	}
-
+func (c Config) Explain(path string) string {
 	explanation := &strings.Builder{}
-	c.explain(explanation, path, sub(c.values.values, strings.Split(path, c.delimiter)), *option)
+	c.explain(explanation, path, maps.Sub(c.values.values, strings.Split(path, c.delimiter)))
 
 	return explanation.String()
 }
 
-func (c Config) explain(explanation *strings.Builder, path string, value any, option explainOptions) {
+func (c Config) explain(explanation *strings.Builder, path string, value any) {
 	if values, ok := value.(map[string]any); ok {
 		for k, v := range values {
-			c.explain(explanation, path+c.delimiter+k, v, option)
+			c.explain(explanation, path+c.delimiter+k, v)
 		}
 
 		return
@@ -197,7 +164,7 @@ func (c Config) explain(explanation *strings.Builder, path string, value any, op
 
 	var loaders []loaderValue
 	for _, provider := range c.values.providers {
-		if v := sub(provider.values, strings.Split(path, c.delimiter)); v != nil {
+		if v := maps.Sub(provider.values, strings.Split(path, c.delimiter)); v != nil {
 			loaders = append(loaders, loaderValue{provider.loader, v})
 		}
 	}
@@ -211,7 +178,7 @@ func (c Config) explain(explanation *strings.Builder, path string, value any, op
 	}
 	explanation.WriteString(path)
 	explanation.WriteString(" has value[")
-	explanation.WriteString(option.valueFormatter(loaders[0].loader, path, loaders[0].value))
+	explanation.WriteString(credential.Blur(path, loaders[0].value))
 	explanation.WriteString("] that is loaded by loader[")
 	explanation.WriteString(fmt.Sprintf("%v", loaders[0].loader))
 	explanation.WriteString("].\n")
@@ -219,7 +186,7 @@ func (c Config) explain(explanation *strings.Builder, path string, value any, op
 		explanation.WriteString("Here are other value(loader)s:\n")
 		for _, loader := range loaders[1:] {
 			explanation.WriteString("  - ")
-			explanation.WriteString(option.valueFormatter(loader.loader, path, loader.value))
+			explanation.WriteString(credential.Blur(path, loader.value))
 			explanation.WriteString("(")
 			explanation.WriteString(fmt.Sprintf("%v", loader.loader))
 			explanation.WriteString(")\n")