+++ provider.go
@@ -3,7 +3,11 @@
 
 package konf
 
-import "context"
+import (
+	"context"
+
+	"github.com/nil-go/konf/internal/maps"
+)
 
 // Loader is the interface that wraps the Load method.
 //
@@ -27,5 +31,5 @@ type Watcher interface {
 //
 // It's used by the loader to check if the configuration has been set by other loaders.
 func (c Config) Exists(path []string) bool {
-	return sub(c.values.values, path) != nil
+	return maps.Sub(c.values.values, path) != nil
 }